project/requirements.txt
Django==3.1.2
django-environ==0.4.5
django-pytest==0.2.0
djangorestframework==3.12.1
factory-boy==3.1.0
Faker==4.14.0
psycopg2==2.8.6
pytest==6.1.1
pytest-django==3.10.0

